+++ components/merkledb/src/tests.rs
@@ -15,8 +15,8 @@
 use exonum_crypto::Hash;
 
 use crate::{
-    Entry, Fork, KeySetIndex, ListIndex, MapIndex, ProofListIndex, ProofMapIndex,
-    SparseListIndex, ValueSetIndex,
+    Entry, Fork, KeySetIndex, ListIndex, MapIndex, ProofListIndex, ProofMapIndex, SparseListIndex,
+    ValueSetIndex,
 };
 
 // This should compile to ensure ?Sized bound on `new_in_family` (see #1024).