+++ exonum/src/api/websocket.rs
@@ -36,8 +36,7 @@ use crate::events::error::into_failure;
 use crate::explorer::{median_precommits_time, TxStatus};
 use crate::messages::{Message as ExonumMessage, ProtocolMessage, RawTransaction, SignedMessage};
 
-use exonum_merkledb::{IndexAccess, Snapshot};
-use exonum_merkledb::proof_list_index::ListProof;
+use exonum_merkledb::{IndexAccess, ListProof, Snapshot};
 
 /// Message, coming from websocket connection.
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]