+++ exonum/src/explorer/mod.rs
@@ -25,6 +25,7 @@ use std::{
     fmt,
     ops::{Index, RangeBounds},
     slice,
+    time::UNIX_EPOCH,
 };
 
 use crate::blockchain::{
@@ -35,9 +36,7 @@ use crate::crypto::{CryptoHash, Hash};
 use crate::helpers::Height;
 use crate::messages::{Precommit, RawTransaction, Signed};
 use chrono::{DateTime, Utc};
-use exonum_merkledb::{ProofVariant, Snapshot};
-use std::time::UNIX_EPOCH;
-use exonum_merkledb::proof_list_index::ListProof;
+use exonum_merkledb::{ListProof, Snapshot};
 
 /// Transaction parsing result.
 type ParseResult = Result<TransactionMessage, failure::Error>;